+++ widgets/linechart/linechart_test.go
@@ -15,6 +15,7 @@
 package linechart
 
 import (
+	"fmt"
 	"image"
 	"math"
 	"testing"
@@ -1634,6 +1635,199 @@ func TestLineChartDraws(t *testing.T) {
 				testdraw.MustBrailleLine(bc, image.Point{9, 5}, image.Point{10, 4})
 				testdraw.MustBrailleLine(bc, image.Point{10, 4}, image.Point{10, 2})
 				testdraw.MustBrailleLine(bc, image.Point{10, 2}, image.Point{11, 0})
+				testbraille.MustCopyTo(bc, c)
+
+				testcanvas.MustApply(c, ft)
+				return ft
+			},
+		},
+		{
+			desc:   "custom Y-axis labels using a value formatter",
+			canvas: image.Rect(0, 0, 20, 10),
+			opts: []Option{
+				YAxisFormattedValues(func(v float64) string {
+					if v == 0 || math.IsNaN(v) {
+						return "∅"
+					}
+					return fmt.Sprintf("%.1fs", v+10)
+				}),
+			},
+			writes: func(lc *LineChart) error {
+				return lc.Series("first", []float64{0, 100})
+			},
+			wantCapacity: 28,
+			want: func(size image.Point) *faketerm.Terminal {
+				ft := faketerm.MustNew(size)
+				c := testcanvas.MustNew(ft.Area())
+
+				// Y and X axis.
+				lines := []draw.HVLine{
+					{Start: image.Point{5, 0}, End: image.Point{5, 8}},
+					{Start: image.Point{5, 8}, End: image.Point{19, 8}},
+				}
+				testdraw.MustHVLines(c, lines)
+
+				// Value labels.
+				testdraw.MustText(c, "∅", image.Point{4, 7})
+				testdraw.MustText(c, "61.7s", image.Point{0, 3})
+				testdraw.MustText(c, "0", image.Point{6, 9})
+				testdraw.MustText(c, "1", image.Point{19, 9})
+
+				// Braille line.
+				graphAr := image.Rect(6, 0, 20, 8)
+				bc := testbraille.MustNew(graphAr)
+				testdraw.MustBrailleLine(bc, image.Point{0, 31}, image.Point{26, 0})
+				testbraille.MustCopyTo(bc, c)
+
+				testcanvas.MustApply(c, ft)
+				return ft
+			},
+		},
+		{
+			desc:   "custom Y-axis labels using a value formatter that returns empty labels",
+			canvas: image.Rect(0, 0, 20, 10),
+			opts: []Option{
+				YAxisFormattedValues(func(v float64) string { return "" }),
+			},
+			writes: func(lc *LineChart) error {
+				return lc.Series("first", []float64{0, 100})
+			},
+			wantCapacity: 38,
+			want: func(size image.Point) *faketerm.Terminal {
+				ft := faketerm.MustNew(size)
+				c := testcanvas.MustNew(ft.Area())
+
+				// Y and X axis.
+				lines := []draw.HVLine{
+					{Start: image.Point{0, 0}, End: image.Point{0, 8}},
+					{Start: image.Point{0, 8}, End: image.Point{19, 8}},
+				}
+				testdraw.MustHVLines(c, lines)
+
+				// Value labels.
+				testdraw.MustText(c, "0", image.Point{1, 9})
+				testdraw.MustText(c, "1", image.Point{19, 9})
+
+				// Braille line.
+				graphAr := image.Rect(1, 0, 20, 8)
+				bc := testbraille.MustNew(graphAr)
+				testdraw.MustBrailleLine(bc, image.Point{0, 31}, image.Point{36, 0})
+				testbraille.MustCopyTo(bc, c)
+
+				testcanvas.MustApply(c, ft)
+				return ft
+			},
+		},
+		{
+			desc:   "custom Y-axis labels using a value formatter that returns very long strings",
+			canvas: image.Rect(0, 0, 20, 10),
+			opts: []Option{
+				YAxisFormattedValues(func(v float64) string {
+					if v == 0 || math.IsNaN(v) {
+						return "0"
+					}
+					return fmt.Sprintf("%.20f", v)
+				}),
+			},
+			writes: func(lc *LineChart) error {
+				return lc.Series("first", []float64{0, 100})
+			},
+			wantCapacity: 2,
+			want: func(size image.Point) *faketerm.Terminal {
+				ft := faketerm.MustNew(size)
+				c := testcanvas.MustNew(ft.Area())
+
+				// Y and X axis.
+				lines := []draw.HVLine{
+					{Start: image.Point{18, 0}, End: image.Point{18, 8}},
+					{Start: image.Point{18, 8}, End: image.Point{19, 8}},
+				}
+				testdraw.MustHVLines(c, lines)
+
+				// Value labels.
+				testdraw.MustText(c, "0", image.Point{17, 7})
+				testdraw.MustText(c, "51.67999999999999…", image.Point{0, 3})
+				testdraw.MustText(c, "0", image.Point{19, 9})
+
+				// Braille line.
+				graphAr := image.Rect(19, 0, 20, 8)
+				bc := testbraille.MustNew(graphAr)
+				testdraw.MustBrailleLine(bc, image.Point{0, 31}, image.Point{1, 0})
+				testbraille.MustCopyTo(bc, c)
+
+				testcanvas.MustApply(c, ft)
+				return ft
+			},
+		},
+		{
+			desc:   "custom Y-axis labels using a value formatter with non printable strings (\\n)",
+			canvas: image.Rect(0, 0, 20, 10),
+			opts: []Option{
+				YAxisFormattedValues(func(v float64) string { return "\n" }),
+			},
+			writes: func(lc *LineChart) error {
+				return lc.Series("first", []float64{0, 100})
+			},
+			wantDrawErr: true,
+		},
+		{
+			desc:   "custom Y-axis labels using a value formatter with non printable strings (\\t)",
+			canvas: image.Rect(0, 0, 20, 10),
+			opts: []Option{
+				YAxisFormattedValues(func(v float64) string { return "\ta" }),
+			},
+			writes: func(lc *LineChart) error {
+				return lc.Series("first", []float64{0, 100})
+			},
+			wantDrawErr: true,
+		},
+		{
+			desc:   "custom Y-axis labels using a value formatter with non printable strings (control characters)",
+			canvas: image.Rect(0, 0, 20, 10),
+			opts: []Option{
+				YAxisFormattedValues(func(v float64) string { return fmt.Sprintf("%ca", 0x007f) }),
+			},
+			writes: func(lc *LineChart) error {
+				return lc.Series("first", []float64{0, 100})
+			},
+			wantDrawErr: true,
+		},
+		{
+			desc:   "custom Y-axis labels using a value formatter that returns unicode strings",
+			canvas: image.Rect(0, 0, 20, 10),
+			opts: []Option{
+				YAxisFormattedValues(func(v float64) string {
+					if v == 0 {
+						return "abc"
+					}
+					return "世世世世"
+				}),
+			},
+			writes: func(lc *LineChart) error {
+				return lc.Series("first", []float64{0, 100})
+			},
+			wantCapacity: 22,
+			want: func(size image.Point) *faketerm.Terminal {
+				ft := faketerm.MustNew(size)
+				c := testcanvas.MustNew(ft.Area())
+
+				// Y and X axis.
+				lines := []draw.HVLine{
+					{Start: image.Point{8, 0}, End: image.Point{8, 8}},
+					{Start: image.Point{8, 8}, End: image.Point{19, 8}},
+				}
+				testdraw.MustHVLines(c, lines)
+
+				// Value labels.
+				testdraw.MustText(c, "abc", image.Point{5, 7})
+				testdraw.MustText(c, "世世世世", image.Point{0, 3})
+				testdraw.MustText(c, "0", image.Point{9, 9})
+				testdraw.MustText(c, "1", image.Point{19, 9})
+
+				// Braille line.
+				graphAr := image.Rect(9, 0, 20, 8)
+				bc := testbraille.MustNew(graphAr)
+				testdraw.MustBrailleLine(bc, image.Point{0, 31}, image.Point{21, 0})
 				testbraille.MustCopyTo(bc, c)
 
 				testcanvas.MustApply(c, ft)